Improves gotoState algorithm for choosing fromState
When using the routing system in combination with concurrent states it lead to problems at transitions. The previous algorithm just looked up current substates and if there were none, it used a null state. Now we walk up the state tree to find the nearest parent/ancestor current state in the tree.